add conform box

+++ src/app/componets/home/home.component.ts
@@ -37,13 +37,15 @@ export class HomeComponent implements OnInit {
 
   deletePost(post: Post) {
     console.log('deleted post : ', post);
-    this.postService.deletePost(post).subscribe(res => {
-      this.posts.forEach((elements, index) => {
-        if (elements.id === post.id) {
-          this.posts.splice(index, 1);
-        }
+    if (confirm('Are you sure you want to delete')) {
+      this.postService.deletePost(post).subscribe(res => {
+        this.posts.forEach((elements, index) => {
+          if (elements.id === post.id) {
+            this.posts.splice(index, 1);
+          }
+        });
       });
-    });
+    }
   }
 
   updatePost(post: Post) {
